Add plain msgpack serialization (#267)
* Add plain msgpack serialization

* Add plain msgpack to more tests & clean up to/from file

* Apply suggestions from code review


* Fix file opening flags in serialization

* For small changes for msgpack(-ext)

+++ qcelemental/tests/test_molecule.py
@@ -219,7 +219,7 @@ def test_to_from_file_simple(tmp_path, dtype, filext):
     )
 
     p = tmp_path / ("water." + filext)
-    benchmol.to_file(p)
+    benchmol.to_file(p, dtype=dtype)
 
     mol = Molecule.from_file(p)
